by claude code

+++ train.py
@@ -245,7 +245,7 @@ elif init_from == 'resume':
 if block_size < model.config.block_size:
     model.crop_block_size(block_size)
     model_args['block_size'] = block_size # so that the checkpoint will have the right value
-model.to(device)
+model.to(device, dtype=ptdtype)
 
 # initialize a GradScaler. If enabled=False scaler is a no-op
 scaler = torch.GradScaler(enabled=(dtype == 'float16'))
